LayoutTests/platform/gtk/accessibility/self-referencing-aria-labelledby-expected.txt

X ?          Y Z
This tests that self-referencing aria-labelledby is exposed correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test.title is 'AXTitle: X Y Z'
PASS test.title is 'AXTitle: X Y Z'
PASS test.title is 'AXTitle: X Y Z'
PASS test.title is 'AXTitle: X Y Z'
PASS test.title is 'AXTitle: X Y Z'
PASS test.title is 'AXTitle: X Y Z'
PASS test.title is 'AXTitle: X Y Z'
PASS successfullyParsed is true

TEST COMPLETE
